Move InputSource extensions to extension package
